2016.5: Integrate with Win32 CNG API
This implementation will be a C++ implementation of 2016.5, initially
executing with multiple CPU threads, but with an eye to provide a Vulkan
compute shader implementation that can run on the GPU.

Step 1 is to integrate with CNG (Cryptography API: Next Generation) to
provide the MD5 hash function on Win32. This commit provides a basic
initialization and test executable, since it's the first time I've
worked with CNG. Compatibility with non-Windows environments needs to
integrate with OpenSSL, which means a portability abstraction across the
two APIs, which will probably form part of step 2, before continuing to
the actual solution implementation.